NN-5684 migrate and sync updates for preprod (#587)
* NN-5684 handle use of back corrupting records

* NN-5684 tolong needs to be behind flag

* NN-5684 remove to long before flag

* NN-5684 tests to confirm accepted as corrupted

* NN-5684 handle last edge cases

* NN-5684 convert active ada exceptions to corrupted accepts

* tidy up

* circle ci playing up

* latest

* testing

* remove

* compile with 21

* set to 19 again circle ci working

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/HearingService.kt
@@ -93,7 +93,7 @@ class HearingService(
     reportedAdjudication.hearings.filter { it.id != hearingToEdit.id }.validateHearingDate(dateTimeOfHearing)
 
     legacySyncService.amendHearing(
-      adjudicationNumber = chargeNumber.toLong(),
+      adjudicationNumber = chargeNumber,
       oicHearingId = hearingToEdit.oicHearingId,
       oicHearingRequest = OicHearingRequest(
         dateTimeOfHearing = dateTimeOfHearing,
@@ -120,7 +120,7 @@ class HearingService(
     val hearingToRemove = reportedAdjudication.getHearing().canDelete()
 
     legacySyncService.deleteHearing(
-      adjudicationNumber = chargeNumber.toLong(),
+      adjudicationNumber = chargeNumber,
       oicHearingId = hearingToRemove.oicHearingId,
     )
