+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/NomisOutcomeServiceTest.kt
@@ -58,7 +58,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
 
       Assertions.assertThatThrownBy {
         nomisOutcomeService.createHearingResultIfApplicable(
-          adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+          adjudicationNumber = reportedAdjudication.chargeNumber,
           hearing = reportedAdjudication.getLatestHearing(),
           outcome = reportedAdjudication.latestOutcome()!!,
         )
@@ -75,7 +75,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.createHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -105,7 +105,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       whenever(legacySyncService.createHearing(any(), any())).thenReturn(123)
 
       val hearingId = nomisOutcomeService.createHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -114,7 +114,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       verify(legacySyncService, atLeastOnce()).quashSanctions(any())
       verify(legacySyncService, atLeastOnce()).createHearing(any(), any())
       verify(legacySyncService, atLeastOnce()).createHearingResult(
-        reportedAdjudication.chargeNumber.toLong(),
+        reportedAdjudication.chargeNumber,
         123,
         OicHearingResultRequest(
           pleaFindingCode = Plea.GUILTY,
@@ -136,7 +136,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       whenever(legacySyncService.createHearing(any(), any())).thenReturn(123)
 
       val hearingId = nomisOutcomeService.createHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -144,7 +144,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       assertThat(hearingId).isNotNull
       verify(legacySyncService, atLeastOnce()).createHearing(any(), any())
       verify(legacySyncService, atLeastOnce()).createHearingResult(
-        reportedAdjudication.chargeNumber.toLong(),
+        reportedAdjudication.chargeNumber,
         123,
         OicHearingResultRequest(
           pleaFindingCode = Plea.NOT_ASKED,
@@ -163,7 +163,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.createHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -182,7 +182,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.createHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -192,7 +192,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
 
       verify(legacySyncService, never()).createHearing(any(), any())
       verify(legacySyncService, atLeastOnce()).createHearingResult(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         oicHearingId = hearing.oicHearingId,
         oicHearingResultRequest = request,
       )
@@ -207,7 +207,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.createHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -229,7 +229,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
 
       Assertions.assertThatThrownBy {
         nomisOutcomeService.amendHearingResultIfApplicable(
-          adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+          adjudicationNumber = reportedAdjudication.chargeNumber,
           hearing = reportedAdjudication.getLatestHearing(),
           outcome = reportedAdjudication.latestOutcome()!!,
         )
@@ -245,7 +245,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
 
       Assertions.assertThatThrownBy {
         nomisOutcomeService.amendHearingResultIfApplicable(
-          adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+          adjudicationNumber = reportedAdjudication.chargeNumber,
           hearing = reportedAdjudication.getLatestHearing(),
           outcome = reportedAdjudication.latestOutcome()!!,
         )
@@ -262,7 +262,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.amendHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -289,14 +289,14 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.amendHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
 
       verify(legacySyncService, never()).createHearing(anyOrNull(), any())
       verify(legacySyncService, atLeastOnce()).amendHearingResult(
-        reportedAdjudication.chargeNumber.toLong(),
+        reportedAdjudication.chargeNumber,
         123L,
         OicHearingResultRequest(
           findingCode = Finding.QUASHED,
@@ -316,14 +316,14 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.amendHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
 
       verify(legacySyncService, never()).createHearing(anyOrNull(), any())
       verify(legacySyncService, atLeastOnce()).amendHearingResult(
-        reportedAdjudication.chargeNumber.toLong(),
+        reportedAdjudication.chargeNumber,
         123L,
         OicHearingResultRequest(
           pleaFindingCode = Plea.NOT_ASKED,
@@ -342,7 +342,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.amendHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -360,7 +360,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.amendHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -369,7 +369,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       val request = createHearingResultRequestForVerify(reportedAdjudication, hearing)
 
       verify(legacySyncService, atLeastOnce()).amendHearingResult(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         oicHearingId = hearing.oicHearingId,
         oicHearingResultRequest = request,
       )
@@ -384,7 +384,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.amendHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -406,7 +406,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
 
       Assertions.assertThatThrownBy {
         nomisOutcomeService.deleteHearingResultIfApplicable(
-          adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+          adjudicationNumber = reportedAdjudication.chargeNumber,
           hearing = reportedAdjudication.getLatestHearing(),
           outcome = reportedAdjudication.latestOutcome()!!,
         )
@@ -423,7 +423,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.deleteHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -452,13 +452,13 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.deleteHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
 
-      verify(legacySyncService, atLeastOnce()).deleteHearing(reportedAdjudication.chargeNumber.toLong(), 123L)
-      verify(legacySyncService, atLeastOnce()).deleteHearingResult(reportedAdjudication.chargeNumber.toLong(), 123L)
+      verify(legacySyncService, atLeastOnce()).deleteHearing(reportedAdjudication.chargeNumber, 123L)
+      verify(legacySyncService, atLeastOnce()).deleteHearingResult(reportedAdjudication.chargeNumber, 123L)
     }
 
     @CsvSource("NOT_PROCEED", "PROSECUTION")
@@ -471,13 +471,13 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.deleteHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
 
-      verify(legacySyncService, atLeastOnce()).deleteHearing(reportedAdjudication.chargeNumber.toLong(), 122L)
-      verify(legacySyncService, atLeastOnce()).deleteHearingResult(reportedAdjudication.chargeNumber.toLong(), 122L)
+      verify(legacySyncService, atLeastOnce()).deleteHearing(reportedAdjudication.chargeNumber, 122L)
+      verify(legacySyncService, atLeastOnce()).deleteHearingResult(reportedAdjudication.chargeNumber, 122L)
     }
 
     @CsvSource("REFER_POLICE", "NOT_PROCEED", "CHARGE_PROVED", "DISMISSED")
@@ -490,7 +490,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.deleteHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
@@ -499,8 +499,8 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
         verify(legacySyncService, atLeastOnce()).deleteSanctions(any())
       }
 
-      verify(legacySyncService, never()).deleteHearing(reportedAdjudication.chargeNumber.toLong(), 100L)
-      verify(legacySyncService, atLeastOnce()).deleteHearingResult(reportedAdjudication.chargeNumber.toLong(), 100L)
+      verify(legacySyncService, never()).deleteHearing(reportedAdjudication.chargeNumber, 100L)
+      verify(legacySyncService, atLeastOnce()).deleteHearingResult(reportedAdjudication.chargeNumber, 100L)
     }
 
     @Test
@@ -512,7 +512,7 @@ class NomisOutcomeServiceTest : ReportedAdjudicationTestBase() {
       }
 
       nomisOutcomeService.deleteHearingResultIfApplicable(
-        adjudicationNumber = reportedAdjudication.chargeNumber.toLong(),
+        adjudicationNumber = reportedAdjudication.chargeNumber,
         hearing = reportedAdjudication.getLatestHearing(),
         outcome = reportedAdjudication.latestOutcome()!!,
       )
